bot/requirements.txt
python-telegram-bot==21.5
httpx==0.27.0
structlog==24.4.0
